refactor: AOP 리팩토링 및 관심사 분리
