+++ backend/src/main/java/com/app/backend/global/aop/AppAspect.java
@@ -3,41 +3,19 @@ package com.app.backend.global.aop;
 import com.app.backend.global.annotation.CustomLock;
 import com.app.backend.global.annotation.CustomPageJsonSerializer;
 import com.app.backend.global.dto.response.ApiResponse;
-import com.app.backend.global.module.CustomPageModule;
-import com.app.backend.global.util.LocalLockManager;
 import com.app.backend.global.util.LockKeyGenerator;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.SerializationFeature;
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import jakarta.annotation.PreDestroy;
+import com.app.backend.global.util.LockManager;
+import com.app.backend.global.util.LockManager.LockWrapper;
+import com.app.backend.global.util.PageUtil;
 import java.lang.reflect.Method;
-import java.util.Optional;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ConcurrentMap;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import java.util.concurrent.ScheduledExecutorService;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.locks.ReentrantLock;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.aspectj.lang.ProceedingJoinPoint;
 import org.aspectj.lang.annotation.Around;
 import org.aspectj.lang.annotation.Aspect;
 import org.aspectj.lang.reflect.MethodSignature;
-import org.redisson.api.RLock;
-import org.redisson.api.RedissonClient;
-import org.redisson.client.RedisConnectionException;
 import org.springframework.data.domain.Page;
 import org.springframework.http.ResponseEntity;
-import org.springframework.transaction.support.TransactionSynchronization;
 import org.springframework.transaction.support.TransactionSynchronizationManager;
 
 @Slf4j
@@ -45,8 +23,6 @@ public class AppAspect {
 
     @Aspect
     public static class PageJsonSerializerAspect {
-
-        private static final ConcurrentMap<String, ObjectMapper> objectMapperMap = new ConcurrentHashMap<>();
 
         @Around("@annotation(com.app.backend.global.annotation.CustomPageJsonSerializer)")
         public Object execute(ProceedingJoinPoint joinPoint) throws Throwable {
@@ -62,68 +38,13 @@ public class AppAspect {
             Object result = joinPoint.proceed();
 
             if (result instanceof Page<?> page)
-                return processPageJson(page, annotation);
+                return PageUtil.processPageJson(page, annotation);
             else if (result instanceof ApiResponse<?> apiResponse)
-                return processApiResponse(apiResponse, annotation);
+                return PageUtil.processApiResponse(apiResponse, annotation);
             else if (result instanceof ResponseEntity<?> responseEntity)
-                return processResponseEntity(responseEntity, annotation);
+                return PageUtil.processResponseEntity(responseEntity, annotation);
             else
                 return result;
-        }
-
-        private ResponseEntity<?> processResponseEntity(final ResponseEntity<?> responseEntity,
-                                                        final CustomPageJsonSerializer annotation) {
-            Object body = responseEntity.getBody();
-            if (body instanceof Page<?> page)
-                return ResponseEntity.ok(processPageJson(page, annotation));
-            else if (body instanceof ApiResponse<?> apiResponse)
-                return ResponseEntity.ok(processApiResponse(apiResponse, annotation));
-            return responseEntity;
-        }
-
-        private ApiResponse<?> processApiResponse(final ApiResponse<?> apiResponse,
-                                                  final CustomPageJsonSerializer annotation) {
-            if (apiResponse.getData() instanceof Page<?> page)
-                return ApiResponse.of(apiResponse.getIsSuccess(),
-                                      apiResponse.getCode(),
-                                      apiResponse.getMessage(),
-                                      processPageJson(page, annotation));
-            return apiResponse;
-        }
-
-        private JsonNode processPageJson(final Page<?> page, final CustomPageJsonSerializer annotation) {
-            String key = generateKey(annotation);
-            ObjectMapper objectMapper = objectMapperMap.computeIfAbsent(key, newKey -> {
-                log.info("새로운 ObjectMapper 생성: {}", newKey);
-                ObjectMapper mapper = new ObjectMapper();
-                mapper.registerModules(new JavaTimeModule(), new CustomPageModule(annotation));
-                mapper.enable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
-                return mapper;
-            });
-
-            try {
-                String json = objectMapper.writeValueAsString(page);
-                return objectMapper.readTree(json);
-            } catch (JsonProcessingException e) {
-                log.error("Json processing exception occurred", e);
-                return null;
-            }
-        }
-
-        private String generateKey(CustomPageJsonSerializer annotation) {
-            return annotation.content() + "_" +
-                   annotation.hasContent() + "_" +
-                   annotation.totalPages() + "_" +
-                   annotation.totalElements() + "_" +
-                   annotation.numberOfElements() + "_" +
-                   annotation.size() + "_" +
-                   annotation.number() + "_" +
-                   annotation.hasPrevious() + "_" +
-                   annotation.hasNext() + "_" +
-                   annotation.isFirst() + "_" +
-                   annotation.isLast() + "_" +
-                   annotation.sort() + "_" +
-                   annotation.empty();
         }
 
     }
@@ -132,21 +53,14 @@ public class AppAspect {
     @RequiredArgsConstructor
     public static class LockAspect {
 
-        private final static int  MAX_UNLOCK_RETRY_COUNT = 3;
-        private final static long RETRY_DELAY            = 100L;
-
-        private final ExecutorService          executorService = Executors.newFixedThreadPool(10);
-        private final ScheduledExecutorService scheduler       = Executors.newScheduledThreadPool(1);
-
-        private final Optional<RedissonClient> redissonClient;
-        private final LocalLockManager         localLockManager;
+        private final LockManager lockManager;
 
         @Around("@annotation(customLock)")
         public Object execute(ProceedingJoinPoint joinPoint, CustomLock customLock) throws Throwable {
             String lockKey = LockKeyGenerator.generateLockKey(joinPoint, customLock.key());
-            LockWrapper lockWrapper = acquireLock(lockKey,
-                                                  customLock.timeUnit().toMillis(customLock.maxWaitTime()),
-                                                  customLock.timeUnit().toMillis(customLock.leaseTime()));
+            LockWrapper lockWrapper = lockManager.acquireLock(lockKey,
+                                                              customLock.timeUnit().toMillis(customLock.maxWaitTime()),
+                                                              customLock.timeUnit().toMillis(customLock.leaseTime()));
 
             if (!lockWrapper.isLocked())
                 throw new RuntimeException("Failed to acquire lock: " + lockKey);
@@ -155,211 +69,15 @@ public class AppAspect {
                 Object result = joinPoint.proceed();
 
                 if (TransactionSynchronizationManager.isActualTransactionActive())
-                    registerLockReleaseAfterTransaction(lockWrapper);
+                    lockManager.registerLockReleaseAfterTransaction(lockWrapper);
                 else
-                    releaseLock(lockWrapper);
+                    lockManager.releaseLock(lockWrapper);
 
                 return result;
             } catch (Throwable e) {
-                releaseLock(lockWrapper);
+                lockManager.releaseLock(lockWrapper);
                 throw e;
             }
-        }
-
-        private LockWrapper acquireLock(final String lockKey, final long maxWaitTime, final long leaseTime) {
-            boolean isRedisAvailable = isRedisRunning();
-
-            RLock         redisLock = isRedisAvailable ? redissonClient.get().getLock(lockKey) : null;
-            ReentrantLock localLock = isRedisAvailable ? null : localLockManager.getLock(lockKey);
-
-            boolean locked = isRedisAvailable
-                             ? tryRedissonLock(redisLock, maxWaitTime, leaseTime)
-                             : tryLocalLock(localLock, maxWaitTime);
-
-            return LockWrapper.of(lockKey, redisLock, localLock, isRedisAvailable && locked, locked);
-        }
-
-        private boolean isRedisRunning() {
-            return redissonClient.map(client -> {
-                try {
-                    client.getKeys().count();
-                    return true;
-                } catch (RedisConnectionException e) {
-                    log.warn("Redis server is not available. Switching to local lock", e);
-                    return false;
-                }
-            }).orElse(false);
-        }
-
-        private boolean tryRedissonLock(final RLock lock, final long maxWaitTime, final long leaseTime) {
-            try {
-                return CompletableFuture.supplyAsync(() -> {
-                    long baseDelay   = 100L;
-                    long elapsedTime = 0L;
-
-                    while (elapsedTime < maxWaitTime) {
-                        try {
-                            if (lock.tryLock(0, leaseTime, TimeUnit.MILLISECONDS))
-                                return true;
-                        } catch (InterruptedException e) {
-                            Thread.currentThread().interrupt();
-                            throw new RuntimeException("Redis lock acquisition interrupted", e);
-                        }
-
-                        log.info("Redis lock acquisition failed, retrying after wait time: {}ms", baseDelay);
-                        sleep(baseDelay);
-
-                        elapsedTime += baseDelay;
-                        baseDelay = Math.min(baseDelay * 2, maxWaitTime - elapsedTime);
-                    }
-                    return false;
-                }, executorService).get();
-            } catch (InterruptedException e) {
-                log.error("Redisson lock acquisition interrupted", e);
-                Thread.currentThread().interrupt();
-                return false;
-            } catch (ExecutionException e) {
-                log.error("Execution exception occurred", e);
-                return false;
-            }
-        }
-
-        private boolean tryLocalLock(final ReentrantLock lock, final long maxWaitTime) {
-            try {
-                return CompletableFuture.supplyAsync(() -> {
-                    long baseDelay   = 100L;
-                    long elapsedTime = 0L;
-
-                    while (elapsedTime < maxWaitTime) {
-                        try {
-                            if (lock.tryLock(0, TimeUnit.MILLISECONDS))
-                                return true;
-                        } catch (InterruptedException e) {
-                            Thread.currentThread().interrupt();
-                            throw new RuntimeException("Local lock acquisition interrupted", e);
-                        }
-
-                        log.info("Local lock acquisition failed, retrying after wait time: {}ms", baseDelay);
-                        sleep(baseDelay);
-
-                        elapsedTime += baseDelay;
-                        baseDelay = Math.min(baseDelay * 2, maxWaitTime - elapsedTime);
-                    }
-                    return false;
-                }, executorService).get();
-            } catch (InterruptedException e) {
-                log.error("Local lock acquisition interrupted", e);
-                Thread.currentThread().interrupt();
-                return false;
-            } catch (ExecutionException e) {
-                log.error("Execution exception occurred", e);
-                return false;
-            }
-        }
-
-        private void releaseLock(final LockWrapper lockWrapper) {
-            if (lockWrapper.usingRedisLock)
-                unlockRedissonLock(lockWrapper.redisLock, 0);
-            else {
-                unlockLocalLock(lockWrapper.localLock);
-                localLockManager.releaseLock(lockWrapper.lockKey, lockWrapper.localLock);
-            }
-        }
-
-        private void unlockRedissonLock(final RLock lock, int retryCount) {
-            if (lock.isLocked() && lock.isHeldByCurrentThread())
-                lock.unlockAsync().whenComplete((unused, throwable) -> {
-                    if (throwable != null) {
-                        log.warn("Failed to unlock redisson lock, retrying {}/{}", retryCount + 1,
-                                 MAX_UNLOCK_RETRY_COUNT);
-                        if (retryCount < MAX_UNLOCK_RETRY_COUNT)
-                            scheduler.schedule(
-                                    () -> unlockRedissonLock(lock, retryCount + 1), RETRY_DELAY, TimeUnit.MILLISECONDS
-                            );
-                        else
-                            forceUnlockRedissonLock(lock);
-                    } else
-                        log.info("Redisson lock successfully unlocked");
-                });
-        }
-
-        private void forceUnlockRedissonLock(final RLock lock) {
-            if (lock.isLocked() && lock.isHeldByCurrentThread()) {
-                lock.forceUnlock();
-                log.warn("Redisson lock forcefully unlocked after max retries");
-            } else
-                log.warn("Skipping force unlock, lock is not held by current thread");
-        }
-
-        private void unlockLocalLock(final ReentrantLock lock) {
-            if (lock.isHeldByCurrentThread())
-                lock.unlock();
-        }
-
-        private void registerLockReleaseAfterTransaction(final LockWrapper lockWrapper) {
-            TransactionSynchronizationManager.registerSynchronization(new TransactionSynchronization() {
-                @Override
-                public void afterCompletion(int status) {
-                    releaseLock(lockWrapper);
-                }
-            });
-        }
-
-        private void sleep(final long millis) {
-            try {
-                Thread.sleep(millis);
-            } catch (InterruptedException e) {
-                Thread.currentThread().interrupt();
-                throw new RuntimeException("Thread sleep interrupted", e);
-            }
-        }
-
-        @PreDestroy
-        private void shutdownExecutors() {
-            log.info("Shutting down executor services...");
-            executorService.shutdown();
-            scheduler.shutdown();
-
-            try {
-                if (!executorService.awaitTermination(5, TimeUnit.SECONDS)) {
-                    log.warn("ExecutorService did not terminate in the specified time.");
-                    executorService.shutdownNow();
-                }
-                if (!scheduler.awaitTermination(5, TimeUnit.SECONDS)) {
-                    log.warn("ScheduledExecutorService did not terminate in the specified time.");
-                    scheduler.shutdownNow();
-                }
-            } catch (InterruptedException e) {
-                log.error("Shutdown interrupted", e);
-                executorService.shutdownNow();
-                scheduler.shutdownNow();
-                Thread.currentThread().interrupt();
-            }
-        }
-
-    }
-
-    @Getter
-    @Builder(access = AccessLevel.PRIVATE)
-    private static class LockWrapper {
-        private final String        lockKey;
-        private final RLock         redisLock;
-        private final ReentrantLock localLock;
-        private final boolean       usingRedisLock;
-        private final boolean       locked;
-
-        private static LockWrapper of(final String lockKey,
-                                      final RLock redisLock,
-                                      final ReentrantLock localLock,
-                                      final boolean usingRedisLock,
-                                      final boolean locked) {
-            return LockWrapper.builder()
-                              .lockKey(lockKey)
-                              .redisLock(redisLock)
-                              .localLock(localLock)
-                              .usingRedisLock(usingRedisLock)
-                              .locked(locked)
-                              .build();
         }
     }
