+++ backend/src/main/java/com/app/backend/global/config/AopConfig.java
@@ -2,9 +2,7 @@ package com.app.backend.global.config;
 
 import com.app.backend.global.aop.AppAspect.LockAspect;
 import com.app.backend.global.aop.AppAspect.PageJsonSerializerAspect;
-import com.app.backend.global.util.LocalLockManager;
-import java.util.Optional;
-import org.redisson.api.RedissonClient;
+import com.app.backend.global.util.LockManager;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
@@ -17,9 +15,8 @@ public class AopConfig {
     }
 
     @Bean
-    public LockAspect redissonLockAspect(final Optional<RedissonClient> redissonClient,
-                                         final LocalLockManager localLockManager) {
-        return new LockAspect(redissonClient, localLockManager);
+    public LockAspect redissonLockAspect(final LockManager lockManager) {
+        return new LockAspect(lockManager);
     }
 
 }
